Changed permits field from int -> long to match method arguments.

+++ java/util/concurrent/Semaphore.java
@@ -5,7 +5,7 @@ package java.util.concurrent;
  * permits.  Each {@link #acquire} blocks if necessary until a permit is
  * available, and then takes it.  Each {@link #release} adds a permit,
  * potentially releasing a blocking acquirer.
- * However, no actual permit objects are used; the <tt>Semaphore</tt> just 
+ * However, no actual permit objects are used; the <tt>Semaphore</tt> just
  * keeps a count of the number available and acts accordingly.
  *
  * <p>Semaphores are used to restrict the number of threads than can
@@ -15,7 +15,7 @@ package java.util.concurrent;
  * class Pool {
  *   private static final MAX_AVAILABLE = 100;
  *   private final Semaphore available = new Semaphore(MAX_AVAILABLE);
- *   
+ *
  *   public Object getItem() throws InterruptedException {
  *     available.acquire();
  *     return getNextAvailableItem();
@@ -31,17 +31,17 @@ package java.util.concurrent;
  *   protected Object[] items = ... whatever kinds of items being managed
  *   protected boolean[] used = new boolean[MAX_AVAILABLE];
  *
- *   protected synchronized Object getNextAvailableItem() { 
+ *   protected synchronized Object getNextAvailableItem() {
  *     for (int i = 0; i < MAX_AVAILABLE; ++i) {
  *       if (!used[i]) {
  *          used[i] = true;
  *          return items[i];
  *       }
  *     }
- *     return null; // not reached 
+ *     return null; // not reached
  *   }
  *
- *   protected synchronized boolean markAsUnused(Object item) { 
+ *   protected synchronized boolean markAsUnused(Object item) {
  *     for (int i = 0; i < MAX_AVAILABLE; ++i) {
  *       if (item == items[i]) {
  *          if (used[i]) {
@@ -57,7 +57,7 @@ package java.util.concurrent;
  *
  * }
  * </pre>
- * <p>Before obtaining an item each thread must acquire a permit from the 
+ * <p>Before obtaining an item each thread must acquire a permit from the
  * semaphore, guaranteeing that an item is available for use. When the
  * thread has finished with the item it is returned back to the pool and
  * a permit is returned to the semaphore, allowing another thread to
@@ -69,10 +69,10 @@ package java.util.concurrent;
  * consistency of the pool itself.
  *
  * <p>A semaphore initialized to one, and which is used such that it only
- * has at most one permit available, can serve as a mutual exclusion lock. 
+ * has at most one permit available, can serve as a mutual exclusion lock.
  * This is more
  * commonly known as a <em>binary semaphore</em>, because it only has two
- * states: one permit available, or zero permits available. 
+ * states: one permit available, or zero permits available.
  * When used in this way, the binary semaphore has the property (unlike many
  * {@link Lock} implementations, that the &quot;lock&quot; can be released by
  * a thread other than the owner (as semaphores have no notion of ownership).
@@ -85,13 +85,13 @@ package java.util.concurrent;
  *
  * @since 1.5
  * @spec JSR-166
- * @revised $Date: 2003/01/20 23:58:52 $
- * @editor $Author: dholmes $
+ * @revised $Date: 2003/01/21 13:29:13 $
+ * @editor $Author: tim $
  *
  */
 public class Semaphore {
 
-    private int permits = 0;
+    private long permits = 0;
 
     /**
      * Construct a <tt>Semaphore</tt> with the given number of
@@ -112,8 +112,8 @@ public class Semaphore {
      * disabled for thread scheduling purposes and lies dormant until
      * one of two things happens:
      * <ul>
-     * <li>Some other thread invokes the {@link #release} method for this 
-     * semaphore and the current thread happens to be chosen as the 
+     * <li>Some other thread invokes the {@link #release} method for this
+     * semaphore and the current thread happens to be chosen as the
      * thread to receive the permit; or
      * <li>Some other thread {@link Thread#interrupt interrupts} the current
      * thread.
@@ -121,17 +121,17 @@ public class Semaphore {
      *
      * <p>If the current thread:
      * <ul>
-     * <li>has its interrupted status set on entry to this method; or 
-     * <li>is {@link Thread#interrupt interrupted} while waiting 
+     * <li>has its interrupted status set on entry to this method; or
+     * <li>is {@link Thread#interrupt interrupted} while waiting
      * for a permit,
      * </ul>
-     * then {@link InterruptedException} is thrown and the current thread's 
-     * interrupted status is cleared. 
+     * then {@link InterruptedException} is thrown and the current thread's
+     * interrupted status is cleared.
      *
      * @throws InterruptedException if the current thread is interrupted
      *
      * @see Thread#interrupt
-     */ 
+     */
     public void acquire() throws InterruptedException {}
 
     /**
@@ -140,7 +140,7 @@ public class Semaphore {
      * with the value <tt>true</tt>,
      * reducing the number of available permits by one.
      *
-     * <p>If no permit is available then this method will return 
+     * <p>If no permit is available then this method will return
      * immediately with the value <tt>false</tt>.
      *
      * @return <tt>true</tt> if a permit was acquired and <tt>false</tt>
@@ -151,18 +151,18 @@ public class Semaphore {
     }
 
     /**
-     * Acquires a permit if one becomes available  within the given waiting 
+     * Acquires a permit if one becomes available  within the given waiting
      * time and the
      * current thread has not been {@link Thread#interrupt interrupted}.
      * <p>Acquires a permit, if one is available and returns immediately,
      * with the value <tt>true</tt>,
      * reducing the number of available permits by one.
      * <p>If no permit is available then
-     * the current thread becomes disabled for thread scheduling 
+     * the current thread becomes disabled for thread scheduling
      * purposes and lies dormant until one of three things happens:
      * <ul>
-     * <li>Some other thread invokes the {@link #release} method for this 
-     * semaphore and the current thread happens to be chosen as the 
+     * <li>Some other thread invokes the {@link #release} method for this
+     * semaphore and the current thread happens to be chosen as the
      * thread to receive the permit; or
      * <li>Some other thread {@link Thread#interrupt interrupts} the current
      * thread; or
@@ -171,15 +171,15 @@ public class Semaphore {
      * <p>If a permit is acquired then the value <tt>true</tt> is returned.
      * <p>If the current thread:
      * <ul>
-     * <li>has its interrupted status set on entry to this method; or 
-     * <li>is {@link Thread#interrupt interrupted} while waiting to acquire 
-     * a permit, 
+     * <li>has its interrupted status set on entry to this method; or
+     * <li>is {@link Thread#interrupt interrupted} while waiting to acquire
+     * a permit,
      * </ul>
-     * then {@link InterruptedException} is thrown and the current thread's 
-     * interrupted status is cleared. 
+     * then {@link InterruptedException} is thrown and the current thread's
+     * interrupted status is cleared.
      * <p>If the specified waiting time elapses then the value <tt>false</tt>
      * is returned.
-     * The given waiting time is a best-effort lower bound. If the time is 
+     * The given waiting time is a best-effort lower bound. If the time is
      * less than or equal to zero, the method will not wait at all.
      *
      * @param timeout the maximum time to wait for a permit
@@ -192,7 +192,7 @@ public class Semaphore {
      * @see Thread#interrupt
      *
      */
-    public boolean tryAcquire(long timeout, TimeUnit granularity) 
+    public boolean tryAcquire(long timeout, TimeUnit granularity)
         throws InterruptedException {
         return false;
     }
